notify user based on preferences
